requirements.txt
langchain==0.2.12
langchain-community==0.2.11
langchain-chroma==0.1.2
chromadb>=0.5.0
posthog>=3.0.0
pydantic-settings>=2.3
python-dotenv==1.0.1
pypdf==4.3.1
gradio==4.20.0
requests>=2.32.3,<3.0.0
ollama==0.6.1
huggingface-hub>=0.21.0
tokenizers>=0.22.0
transformers>=4.40.0
sentence-transformers>=5.0.0
rank-bm25>=0.2.2
fastapi>=0.104.0
uvicorn>=0.24.0
faster-whisper>=1.0.0
python-multipart>=0.0.6